requirements.txt
djangorestframework
djoser
django-cors-headers
djongo
pyts
pymongo==3.12.3